review: suppress overlay for disabled rows, make the menu builder an owning type
A row's (or ancestor's) .disabled(true) reaches the overlay as SwiftUI's
isEnabled environment; the representable now mirrors it into the view
and a disabled row offers no context menu on mouse or accessibility
paths, matching native SwiftUI. The builder becomes an instance struct
with constructor-injected dispatch (owned entry point, not an ambient
free function), pure helpers stay file-scope private, and the nested-
overlay search moves to file scope per Aziz policy.
